Refactor packet and packet-list components
Extract the packet value table into a separate component
and use it in both packet and packet-list.
