+++ yamcs-web/src/main/webapp/projects/webapp-sdk/src/lib/components/detail-pane/detail-pane.component.css
@@ -21,7 +21,7 @@
 }
 
 .detail-content.hidden {
-  opacity: 0;
+  visibility: hidden;
 }
 
 .toggle-button {
