+++ yamcs-web/src/main/webapp/projects/webapp/src/app/telemetry/packets/packet-detail-table/packet-detail-table.component.html
@@ -1,0 +1,205 @@
+<table
+  mat-table
+  [dataSource]="dataSource"
+  class="ya-data-table expand"
+  multiTemplateDataRows>
+  <ng-container matColumnDef="icon">
+    <th mat-header-cell *matHeaderCellDef width="1"></th>
+    <td mat-cell *matCellDef="let node" class="vtop">
+      @if (node.type === "CONTAINER") {
+        <ya-icon-action
+          [icon]="node.expanded ? 'expand_more' : 'chevron_right'"
+          [padding]="false" />
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="location">
+    <th mat-header-cell *matHeaderCellDef class="no-left-padding" width="1">Loc</th>
+    <td mat-cell *matCellDef="let node" class="vtop right no-left-padding">
+      @if (
+        node.type === "SIMPLE_PARAMETER" ||
+        node.type === "AGGREGATE_PARAMETER" ||
+        node.type === "ARRAY_PARAMETER"
+      ) {
+        {{ node.location }}
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="size">
+    <th mat-header-cell *matHeaderCellDef width="1">Bits</th>
+    <td mat-cell *matCellDef="let node" class="vtop right">
+      @if (
+        node.type === "SIMPLE_PARAMETER" ||
+        node.type === "AGGREGATE_PARAMETER" ||
+        node.type === "ARRAY_PARAMETER"
+      ) {
+        {{ node.size }}
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="expand-aggray">
+    <th mat-header-cell *matHeaderCellDef width="1" class="icon"></th>
+    <td mat-cell *matCellDef="let node" class="icon vtop">
+      @if (node.type === "AGGREGATE_PARAMETER" || node.type === "ARRAY_PARAMETER") {
+        <ya-icon-action [icon]="node.expanded ? 'expand_more' : 'chevron_right'" />
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="entry">
+    <th mat-header-cell *matHeaderCellDef class="entry">Entry</th>
+    <td mat-cell *matCellDef="let node" class="entry vtop">
+      @if (
+        node.type === "SIMPLE_PARAMETER" ||
+        node.type === "AGGREGATE_PARAMETER" ||
+        node.type === "ARRAY_PARAMETER"
+      ) {
+        <a
+          [routerLink]="'/telemetry/parameters' + node.parameter.qualifiedName"
+          [queryParams]="{ c: yamcs.context }"
+          class="ya-link">
+          {{ node.parameter.qualifiedName }}
+        </a>
+      }
+      @if (node.type === "SIMPLE_VALUE") {
+        @for (x of [].constructor(node.depth); track x; let last = $last) {
+          <mat-icon class="icon12" style="visibility: hidden">chevron_right</mat-icon>
+        }
+        <a
+          [routerLink]="
+            '/telemetry/parameters' + node.parameter.qualifiedName + node.offset
+          "
+          [queryParams]="{ c: yamcs.context }"
+          class="ya-link">
+          {{ node.name }}
+        </a>
+      }
+      @if (node.type === "AGGREGATE_VALUE") {
+        @for (x of [].constructor(node.depth); track x; let last = $last) {
+          <mat-icon class="icon12" [style.visibility]="last ? 'visible' : 'hidden'">
+            {{ node.expanded ? "expand_more" : "chevron_right" }}
+          </mat-icon>
+        }
+        <a
+          [routerLink]="
+            '/telemetry/parameters' + node.parameter.qualifiedName + node.offset
+          "
+          [queryParams]="{ c: yamcs.context }"
+          class="ya-link">
+          {{ node.name }}
+        </a>
+      }
+      @if (node.type === "ARRAY_VALUE") {
+        @for (x of [].constructor(node.depth); track x; let last = $last) {
+          <mat-icon class="icon12" [style.visibility]="last ? 'visible' : 'hidden'">
+            {{ node.expanded ? "expand_more" : "chevron_right" }}
+          </mat-icon>
+        }
+        <a
+          [routerLink]="
+            '/telemetry/parameters' + node.parameter.qualifiedName + node.offset
+          "
+          [queryParams]="{ c: yamcs.context }"
+          class="ya-link">
+          {{ node.name }}
+        </a>
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="type">
+    <th mat-header-cell *matHeaderCellDef>Type</th>
+    <td mat-cell *matCellDef="let node" class="mono vtop">
+      @if (node.type === "CONTAINER") {
+        container
+      }
+      @if (node.type === "SIMPLE_PARAMETER") {
+        {{ node.parameter.type.engType }}
+      }
+      @if (node.type === "AGGREGATE_PARAMETER") {
+        {{ node.parameter.type.engType }}
+      }
+      @if (node.type === "ARRAY_PARAMETER") {
+        {{
+          fillTypeWithValueDimension(node.parameter.type.engType, node.engValue?.arrayValue)
+        }}
+      }
+      @if (node.type === "SIMPLE_VALUE") {
+        {{ node.parameterType.engType }}
+      }
+      @if (node.type === "AGGREGATE_VALUE") {
+        {{ node.parameterType.engType }}
+      }
+      @if (node.type === "ARRAY_VALUE") {
+        {{
+          fillTypeWithValueDimension(node.parameterType.engType, node.engValue?.arrayValue)
+        }}
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="rawValue">
+    <th mat-header-cell *matHeaderCellDef>Raw value</th>
+    <td mat-cell *matCellDef="let node" class="vtop wrap200">
+      @if (node.rawValue; as value) {
+        @if (node.type === "SIMPLE_PARAMETER") {
+          <ya-value [value]="value" />
+        }
+        @if (node.type === "SIMPLE_VALUE") {
+          <ya-value [value]="value" />
+        }
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="engValue">
+    <th mat-header-cell *matHeaderCellDef>Engineering value</th>
+    <td mat-cell *matCellDef="let node" class="vtop wrap200">
+      @if (node.engValue; as value) {
+        @if (node.type === "SIMPLE_PARAMETER") {
+          <ya-value [value]="value" />
+          @if (node.parameter.type?.unitSet; as unitSet) {
+            {{ unitSet | units }}
+          }
+        }
+        @if (node.type === "SIMPLE_VALUE") {
+          <ya-value [value]="value" />
+          @if (node.parameterType.unitSet; as unitSet) {
+            {{ unitSet | units }}
+          }
+        }
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="containerName">
+    <th mat-header-cell *matHeaderCellDef class="no-left-padding"></th>
+    <td mat-cell *matCellDef="let node" [attr.colspan]="4" class="no-left-padding">
+      @if (node.type === "CONTAINER") {
+        {{ node.container.qualifiedName }}
+      }
+    </td>
+  </ng-container>
+  <ng-container matColumnDef="actions">
+    <th mat-header-cell *matHeaderCellDef class="expand"></th>
+    <td mat-cell *matCellDef="let node" class="icon">
+      @if (
+        node.type === "SIMPLE_PARAMETER" ||
+        node.type === "AGGREGATE_PARAMETER" ||
+        node.type === "ARRAY_PARAMETER"
+      ) {
+        <ya-more [padding]="false">
+          <button mat-menu-item (click)="selectBitRange(node)">Select in hex view</button>
+        </ya-more>
+      }
+    </td>
+  </ng-container>
+  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+  <tr
+    mat-row
+    *matRowDef="let node; columns: containerColumns; when: isContainer"
+    class="expandable container"
+    (click)="toggleRow(node)"></tr>
+  <tr
+    mat-row
+    *matRowDef="let node; columns: displayedColumns; when: isNoContainer"
+    [class.expandable]="isExpandable(node)"
+    (mouseover)="highlightBitRange(node)"
+    (mouseout)="clearHighlightedBitRange()"
+    (click)="toggleRow(node)"></tr>
+</table>
