+++ yamcs-web/src/main/webapp/projects/webapp/src/app/telemetry/packets/packet-list/packet-list.component.html
@@ -59,6 +59,15 @@
         <div class="block" style="margin-top: 4px">
           <app-hex #hexdump [base64String]="detailPacket.packet" />
         </div>
+        <div class="block-title">Values</div>
+        <div class="block">
+          <app-packet-detail-table
+            [packet]="detailPacket"
+            (highlightBitRangeEvent)="onHighlightBitRange($event)"
+            (clearHighlightedBitRangeEvent)="onClearHighlightedBitRange()"
+            (selectBitRangeEvent)="onSelectBitRange($event)">
+          </app-packet-detail-table>
+        </div>
       </div>
     } @else {
       <ya-detail-toolbar>Select a packet</ya-detail-toolbar>
